Move `retry` package to `internal/retry`
